feat: first working sequential simulation!
TODO: use and document appropiate units!
 Currently the pixel distance is uses in Combination with an arbitrary mass...

+++ src/physics.rs
@@ -1,1 +1,1 @@
-pub const GRAVITY: f32 = 9.81;
+pub const GRAVITY: f64 = 9.81;
